+++ R/link.R
@@ -282,12 +282,25 @@ href_article <- function(article, package = NULL) {
   }
 
   base_url <- remote_package_article_url(package)
-  if (is.null(base_url)) {
-    paste0("https://cran.rstudio.com/web/packages/", package, "/vignettes/", path)
-  } else {
+  if (!is.null(base_url)) {
     paste0(base_url, "/", path)
+  } else if (is_bioc_pkg(package)) {
+    paste0("https://bioconductor.org/packages/release/bioc/vignettes/", package, "/inst/doc/", path)
+  } else {
+    paste0("https://cran.rstudio.com/web/packages/", package, "/vignettes/", path)
   }
 }
+
+# Returns NA if package is not installed.
+# Returns TRUE if `package` is from Bioconductor, FALSE otherwise
+is_bioc_pkg <- function(package) {
+  if (!rlang::is_installed(package)) {
+    return(FALSE)
+  }
+  biocviews <- utils::packageDescription(package, fields = "biocViews")
+  !is.na(biocviews) && biocviews != ""
+}
+
 
 # Try to figure out package name from attached packages
 find_vignette_package <- function(x) {