+++ tests/testthat/test-link.R
@@ -164,6 +164,15 @@ test_that("can link to local articles", {
   expect_equal(href_expr_(vignette("y")), NA_character_)
 })
 
+test_that("can link to bioconductor vignettes", {
+  skip_if_not_installed("MassSpecWavelet")
+
+  expect_equal(
+    href_expr_(vignette("MassSpecWavelet", "MassSpecWavelet")),
+    "https://bioconductor.org/packages/release/bioc/vignettes/MassSpecWavelet/inst/doc/MassSpecWavelet.html"
+  )
+})
+
 test_that("can link to remote articles", {
   skip_on_cran()