+++ NEWS.md
@@ -19,6 +19,8 @@
 
 * Link to correct topic with `::()` and `utils::help()` (@IndrajeetPatil, #131).
 
+* Generate correct link for Bioconductor vignettes (@zeehio, #145)
+
 # downlit 0.4.0
 
 ## Syntax highlighting